resource/{{cookiecutter.script_name}}/requirements.txt
cloudshell-automation-api
cloudshell-logging
cloudshell-sandbox-reporter
cloudshell-cli
